* wtar is created as .wtar.aa even if not big enough to be split * wtars are not created if there is an existing wtar and contents are identical, if they are created do_wtar will erase previous .wtar files * InstlAdmin.stage2svn_for_folder: wtars are not copied or added during stage2svn if there is an existing wtar and contents are identical. * above rule applies even when new file is .wtar.aa and old is .wtar * InstlAdmin.prepare_limit_list renamed prepare_list_of_dirs_to_work_on * new function misc_utils.get_wtar_total_checksum
instl 1.5.3.0

+++ pyinstl/platformSpecificHelper_Mac.py
@@ -267,12 +267,17 @@ class PlatformSpecificHelperMac(PlatformSpecificHelperBase):
         the_split_func = ("""
 split_file()
 {
-    file_size=$(stat -f %z "$1")
-    if [ "$(MIN_FILE_SIZE_TO_WTAR)" -lt "$file_size" ]
+    if [ -f "$1" ]
     then
-        let "part_size=($file_size / (($file_size / $(MIN_FILE_SIZE_TO_WTAR)) + ($file_size % $(MIN_FILE_SIZE_TO_WTAR) > 0 ? 1 : 0)))+1"
-        split -a 2 -b $part_size "$1" "$1."
-        rm -fr "$1"
+        file_size=$(stat -f %z "$1")
+        if [ "$(MIN_FILE_SIZE_TO_WTAR)" -lt "$file_size" ]
+        then
+            let "part_size=($file_size / (($file_size / $(MIN_FILE_SIZE_TO_WTAR)) + ($file_size % $(MIN_FILE_SIZE_TO_WTAR) > 0 ? 1 : 0)))+1"
+            split -a 2 -b $part_size "$1" "$1."
+            rm -fr "$1"
+        else
+            mv "$1" "$1.aa"
+        fi
     fi
 }
 """)
